feat(ToFSensor) : implement ToFsensor driver

+++ include/interfaces/ToFSensorInterface.h
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <inttypes.h>
+
 class ToFSensorInterface
 {
     public:
         virtual void powerOn() = 0;
         virtual void powerOff() = 0;
         virtual bool init() = 0;
-        virtual int getDistance() = 0;
+        virtual uint16_t getDistance() = 0;
 };
